% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataChord.R
\docType{data}
\name{CHORD}
\alias{CHORD}
\title{CHORD (Classifier for Homologous Recombination Deficiency)}
\format{An object of class \code{randomForest} of length 18.}
\usage{
data(CHORD)
}
\description{
A random forest predictor of homologous recombination deficiency
}
\details{
Version 38: 38_HMF_pkgMutSigs_germLt3-somEq0-lohEq0_greylistIncl_multiclass_snvContext_svContext_allSigsRel_scaleGridSearch_featPreFilt_boruta_newAnn
}
\keyword{datasets}
