% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2SingleCellExperiment_INPUT.R
\name{creatSCE}
\alias{creatSCE}
\title{translate seurat obj to SCE}
\usage{
creatSCE(seu)
}
\arguments{
\item{seu}{the seurat object}
}
\description{
simulation training set is generated from quality singlet data after filter any doublets detected by any method
}
