% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSigsCHORD.R
\name{extractSigsChord}
\alias{extractSigsChord}
\title{Extract signatures in the format compatible with CHORD}
\usage{
extractSigsChord(vcf.snv, vcf.indel, vcf.sv, sample.name = "sample",
  sv.caller = "manta")
}
\arguments{
\item{vcf.snv}{Path to the vcf file containing SNVs}

\item{vcf.indel}{Path to the vcf file containing indels}

\item{vcf.sv}{Path to the vcf file containing SVs (Manta vcf)}

\item{sample.name}{The name of the sample as a character. Defaults to 'sample' if none is provided.}
}
\value{
A 1-row data frame containing the mutational signature contributions
}
\description{
This function is a wrapper for the 3 functions from mutSigExtractor:
extractSigsSnv(), extractSigsIndel(), extractSigsSv(). Some post-processing is done to produce
compatible input for CHORD
}
\examples{
sigs <- extractSigsChord(
   vcf.snv = '/Users/lnguyen/hpc/cog_bioinf/cuppen/project_data/HMF_data/DR010/analysis/Arne/DR-10_update/scripts/SNV2//XXXXXXXX.vcf.gz',
   vcf.indel = '/Users/lnguyen/hpc/cog_bioinf/cuppen/project_data/HMF_data/DR010/analysis/Arne/DR-10_update/scripts/INDEL2//XXXXXXXX.vcf.gz',
   vcf.sv = '/Users/lnguyen/hpc/cog_bioinf/cuppen/project_data/HMF_data/DR010-update/data//170217_HMFregXXXXXXXX/XXXXXXXX.vcf.gz',
   sample.name = 'XXXXXXXX'
)
rf_model <- readRDS('/Users/lnguyen/hpc/cog_bioinf/cuppen/project_data/Luan_PCAGW/scripts/mltoolkit/example/data/rf_hrd_predict.rds')
predict(rf_model, sigs, type = 'prob')
}
