% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureByClassCor.R
\name{featureByClassCor}
\alias{featureByClassCor}
\title{T-test between positive vs. negative class}
\usage{
featureByClassCor(
  df,
  feature,
  colname.response = "response",
  p.value.p = 0.01,
  p.value.n = 0.01,
  p.responses = c("BRCA1", "BRCA2"),
  n.response = "none"
)
}
\arguments{
\item{df}{A dataframe containing the features and a response (class label) column}

\item{feature}{Name of the feature column}

\item{colname.response}{Name of the response column}

\item{p.value.p}{Positively correlated features with p-values higher than this are removed}

\item{p.value.n}{Negatively correlated features with p-values higher than this are removed}

\item{p.responses}{Name(s) of the positive class}

\item{n.response}{Name of the negative class}
}
\value{
A list containing a character vector indicating which features are to be kept; and the
t-test results
}
\description{
T-test between positive vs. negative class
}
