% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkdir.R
\name{mkdir}
\alias{mkdir}
\title{Imitates the bash mkdir -p}
\usage{
mkdir(path, recursive = T, ...)
}
\arguments{
\item{path}{Path}

\item{recursive}{Analogous to -p option}

\item{...}{}
}
\description{
Imitates the bash mkdir -p
}
