% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1overkillDB.R
\name{overkillDB2}
\alias{overkillDB2}
\title{overkill the real doublets}
\usage{
overkillDB2(
  seu,
  sce,
  doubletrate,
  seed = 1,
  k = 20,
  overkill = T,
  overkillrate = 1
)
}
\arguments{
\item{seu}{the seurat object}

\item{sce}{the SingleCellExperiment object}

\item{seed}{random seed}

\item{k}{k-means param k}

\item{overkill}{weather use overkill}

\item{overkillrate}{remove the top ?\% doublet-liked cells of any methods' results.}
}
\description{
simulation training set is generated from quality singlet data after filter any doublets detected by any method
}
