% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.chord.R
\name{randomForest.chord}
\alias{randomForest.chord}
\title{Default randomForest() settings for training CHORD}
\usage{
randomForest.chord(df = hmf_training_set, colname.response = "response", ...)
}
\arguments{
\item{df}{A dataframe containing the features and a response (class label) column}

\item{colname.response}{Name of the response column}

\item{...}{Other arguments that can be passed to randomForest()}
}
\value{
A random forest object
}
\description{
Default randomForest() settings for training CHORD
}
