% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForestCv.chord.R
\name{randomForestCv.chord}
\alias{randomForestCv.chord}
\title{Cross-validation wrapper for randomForest.chord()}
\usage{
randomForestCv.chord(df, colname.response = "response", cv.folds = 10,
  cv.repeats = 1, cv.verbose = T, cv.keep.forests = T, ...)
}
\arguments{
\item{df}{A dataframe containing the features and a response (class label) column}

\item{colname.response}{Name of the response column}

\item{cv.folds}{Number of cross-validation folds}

\item{cv.repeats}{Number of times to repeat the cross-validation}

\item{cv.verbose}{Print messages?}

\item{cv.keep.forests}{Keep the random forests or only keep the predictions?}

\item{...}{Other arguments that can be passed to randomForest.chord()}
}
\value{
A list of random forests and/or the predictions
}
\description{
Cross-validation wrapper for randomForest.chord()
}
