% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForestCvNested.chord.R
\name{randomForestCvNested.chord}
\alias{randomForestCvNested.chord}
\title{Outer cross-validation wrapper for  the training and testing procedure}
\usage{
randomForestCvNested.chord(
  df = hmf_training_set,
  colname.response = "response",
  k.outer = 10,
  parent.export.dir = NULL,
  rfcv.nested.seed = NULL,
  rfcv.nested.verbose = T,
  ...
)
}
\arguments{
\item{df}{A dataframe containing the features and a response (class label) column}

\item{colname.response}{Name of the response column}

\item{k.outer}{Number of outer cross-validation folds}

\item{parent.export.dir}{Dir to output results}

\item{rfcv.nested.seed}{Set seed}

\item{rfcv.nested.verbose}{Print messages?}

\item{...}{Other arguments that can be passed to randomForestTrainTest.chord()}
}
\description{
Outer cross-validation wrapper for  the training and testing procedure
}
