% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForestTrainTest.chord.R
\name{randomForestTrainTest.chord}
\alias{randomForestTrainTest.chord}
\title{Random forest training and testing procedure}
\usage{
randomForestTrainTest.chord(
  train,
  test = NULL,
  colname.response = "response",
  do.feature.prefilter = T,
  feature.prefilter.p.value.n = 1,
  feature.prefilter.p.value.p = 1,
  do.boruta = T,
  do.balance = T,
  balance.aucpr.tol = NULL,
  inner.cv.repeats = 10,
  train.test.seed = NULL,
  train.test.verbose = T,
  export.dir = NULL,
  return.output = F,
  ...
)
}
\arguments{
\item{train}{Training set}

\item{test}{Test set}

\item{colname.response}{Name of the response column}

\item{do.feature.prefilter}{Do t-test filtering?}

\item{feature.prefilter.p.value.n}{Negatively correlated features with p-values higher than this 
are removed}

\item{feature.prefilter.p.value.p}{Positively correlated features with p-values higher than this 
are removed}

\item{do.boruta}{Do boruta feature selection?}

\item{do.balance}{Do class scaling?}

\item{inner.cv.repeats}{Number of inner cross-validation repeats}

\item{train.test.seed}{Set seed}

\item{train.test.verbose}{Print messages?}

\item{export.dir}{Dir to output results}

\item{return.output}{Return output in R?}

\item{...}{Other arguments that can be passed to randomForest.chord()}
}
\value{
A random forest and the test predictions
}
\description{
Random forest training and testing procedure
}
