% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runrunrun.R
\name{chord}
\alias{chord}
\title{Chord}
\usage{
chord(
  seu = NA,
  sce = NA,
  doubletrate = NA,
  mfinal = 40,
  k = 20,
  overkill = T,
  overkillrate = 1,
  outname = "out",
  seed = 1,
  addmethods1 = NA,
  addmethods2 = NA,
  method = "gbm"
)
}
\arguments{
\item{seu}{the input sce object}

\item{mfinal}{an integer, the number of iterations for which boosting is run or the number of trees to use. Defaults to mfinal=40 iterations.}

\item{k}{an integer,k-means param k}

\item{overkill}{if True,use overkill}

\item{overkillrate}{an integer,remove the top ?\% doublet-liked cells of any methods' results.(0-1)}

\item{outname}{The prefix of the output file}

\item{seed}{an integer, random seed}

\item{addmethods1}{the table merged with other method's scores1}

\item{addmethods2}{the table merged with other method's scores2}

\item{method}{boosting method}
}
\description{
remove doublet with scds,bcds,DoubletFinder,doubletCells.
}
\examples{
chord(seu=NA,doubletrate=NA,mfinal=40,k=20,overkill=T,overkillrate=1,outname="out",seed=1,method="gbm")
}
