% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runrunrun.R
\name{chord}
\alias{chord}
\title{Chord}
\usage{
chord(
  seu = NA,
  sce = NA,
  doubletrate = NA,
  mfinal = 40,
  k = 20,
  method = "gbm",
  overkill = T,
  overkillrate = 1,
  outname = "out",
  seed = 1,
  addmethods1 = NA,
  addmethods2 = NA,
  overkilllist = NA,
  adddoublt = NA,
  cxds.ntop = NA,
  cxds.binThresh = NA,
  bcds.ntop = NA,
  bcds.srat = NA,
  dbf.PCs = 1:10,
  dbf.pN = 0.25,
  dbf.pK = NA
)
}
\arguments{
\item{seu}{the input sce object}

\item{mfinal}{an integer, the number of iterations for which boosting is run or the number of trees to use. Defaults to mfinal=40 iterations.(only works when method="adaboost")}

\item{k}{an integer,k-means param k}

\item{method}{the boost method ("adaboost" or "gbm")}

\item{overkill}{if True,use overkill}

\item{overkillrate}{an integer,remove the top ?\% doublet-liked cells of any methods' results.(0-1)}

\item{outname}{The prefix of the output file}

\item{seed}{an integer, random seed}

\item{addmethods1}{the table merged with other method's scores1}

\item{addmethods2}{the table merged with other method's scores2}

\item{overkilllist}{a vector of cells to be remove in overkill}

\item{adddoublt}{doubletrate of cells to be simulate}

\item{cxds.ntop}{integer, indimessageing number of top variance genes to consider. Default: 500}

\item{cxds.binThresh}{integer, minimum counts to consider a gene "present" in a cell. Default: 0}

\item{bcds.ntop}{integer, indicating number of top variance genes to consider. Default: 500}

\item{bcds.srat}{numeric, indicating ratio between orginal number of "cells" and simulated doublets; Default: 1}

\item{dbf.PCs}{Number of statistically-significant principal components (e.g., as estimated from PC elbow plot); Default: 1:10}

\item{dbf.pN}{The number of generated artificial doublets, expressed as a proportion of the merged real-artificial data. Default is set to 0.25, based on observation that DoubletFinder performance is largely pN-invariant (see McGinnis, Murrow and Gartner 2019, Cell Systems).}

\item{dbf.pK}{The PC neighborhood size used to compute pANN, expressed as a proportion of the merged real-artificial data. No default is set, as pK should be adjusted for each scRNA-seq dataset. Optimal pK values can be determined using mean-variance-normalized bimodality coefficient.}
}
\description{
remove doublet with scds,bcds,DoubletFinder.
}
\examples{
chord(seu=NA,doubletrate=NA,k=20,overkill=T,overkillrate=1,outname="out",seed=1)
}
