% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataChord.R
\docType{data}
\name{CHORD}
\alias{CHORD}
\title{CHORD (Classifier for Homologous Recombination Deficiency) v2}
\format{
An object of class \code{randomForest} of length 21.
}
\usage{
data(CHORD)
}
\description{
A random forest predictor of homologous recombination deficiency
}
\keyword{datasets}
