% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataChord.R
\docType{data}
\name{CHORD}
\alias{CHORD}
\title{CHORD (Classifier for Homologous Recombination Deficiency)}
\format{An object of class \code{randomForest} of length 18.}
\usage{
data(CHORD)
}
\description{
A random forest predictor of homologous recombination deficiency
}
\details{
Version 60: 60_snvContext_svContext_allSigsRel_noSuppEvidence_customBlacklist3_noBoruta
}
\keyword{datasets}
