% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataChord.R
\docType{data}
\name{CHORD}
\alias{CHORD}
\title{CHORD (Classifier for Homologous Recombination Deficiency)}
\format{An object of class \code{randomForest} of length 19.}
\usage{
data(CHORD)
}
\description{
A random forest predictor of homologous recombination deficiency
}
\details{
Version 17: 17_HMF_pkgMutSigs_germLt3-somEq0-lohEq0_delmhOnly_noNegCorFeat_greylistIncl_brca1Scale1point5x_noneScale0point33x_multiclass_featSel
}
\keyword{datasets}
