% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chordPredict.R
\name{chordPredict}
\alias{chordPredict}
\title{Predict the probability of homogolous recombination deficiency using mutational signatures}
\usage{
chordPredict(newdata, raw.output = F)
}
\arguments{
\item{newdata}{A data frame or matrix containing new data}

\item{raw.output}{If TRUE, the probabilities of 'BRCA1', 'BRCA2' and 'none' (no predicted
mutation) will also be returned}
}
\value{
A dataframe containing the prediction probabilities
}
\description{
A wrapper for predict.randomForest() from the randomForest package
}
