% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSigsChord.R
\name{extractSigsChord}
\alias{extractSigsChord}
\title{Extract mutation contexts in the format compatible with CHORD}
\usage{
extractSigsChord(vcf.snv, vcf.indel = vcf.snv, vcf.sv,
  sample.name = "sample", sv.caller = "gridss", output.path = NULL,
  verbose = T)
}
\arguments{
\item{vcf.snv}{Path to the vcf file containing SNVs}

\item{vcf.indel}{Path to the vcf file containing indels. By default vcf.indel==vcf.snv}

\item{vcf.sv}{Path to the vcf file containing SVs}

\item{sample.name}{The name of the sample as a character. Defaults to 'sample' if none is 
provided.}

\item{sv.caller}{SV vcfs are not standardized and therefore need to be parsed differently
depending on the caller. Currently supports 'manta' or 'gridss'.}

\item{output.path}{If a path is specified, the output is written to this path.}

\item{verbose}{Whether to print progress messages}
}
\value{
A 1-row data frame containing the mutational signature contributions
}
\description{
This function is a wrapper for the 3 functions from mutSigExtractor:
extractSigsSnv(), extractSigsIndel(), extractSigsSv(). Some post-processing is done to produce
compatible input for CHORD
}
